added status for return value in controllers | validation in request dtos

+++ src/main/java/hakaton/eventservice/service/impl/EventServiceImpl.java
@@ -1,13 +1,12 @@
 package hakaton.eventservice.service.impl;
 
-import hakaton.eventservice.exception.NoEntityException;
+import java.util.List;
+import hakaton.eventservice.exception.NoEntityFoundException;
 import hakaton.eventservice.model.Event;
 import hakaton.eventservice.repository.EventRepository;
 import hakaton.eventservice.service.EventService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @AllArgsConstructor
@@ -22,7 +21,7 @@ public class EventServiceImpl implements EventService {
     @Override
     public Event getById(Long id) {
         return eventRepository.findById(id).orElseThrow(() -> {
-            throw new NoEntityException("Event with id: " + id + " doesn't exist");
+            throw new NoEntityFoundException("Event with id: " + id + " doesn't exist");
         });
     }
 
@@ -33,13 +32,16 @@ public class EventServiceImpl implements EventService {
 
     @Override
     public void deleteById(Long id) {
+        if (eventRepository.existsById(id)) {
+            throw new NoEntityFoundException("Event with id: " + id + " doesn't exist");
+        }
         eventRepository.deleteById(id);
     }
 
     @Override
     public Event update(Event event) {
         if (!eventRepository.existsById(event.getId())) {
-            throw new NoEntityException("Event: " + event + " doesn't exist");
+            throw new NoEntityFoundException("Event: " + event + " doesn't exist");
         }
         return eventRepository.save(event);
     }
