+++ src/main/java/hakaton/eventservice/service/impl/EventStatusServiceImpl.java
@@ -1,13 +1,12 @@
 package hakaton.eventservice.service.impl;
 
-import hakaton.eventservice.exception.NoEntityException;
+import java.util.List;
+import hakaton.eventservice.exception.NoEntityFoundException;
 import hakaton.eventservice.model.EventStatus;
 import hakaton.eventservice.repository.EventStatusRepository;
 import hakaton.eventservice.service.EventStatusService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @AllArgsConstructor
@@ -22,7 +21,7 @@ public class EventStatusServiceImpl implements EventStatusService {
     @Override
     public EventStatus getById(Long id) {
         return eventStatusRepository.findById(id).orElseThrow(() -> {
-            throw new NoEntityException("EventStatus with id: " + id + " doesn't exist");
+            throw new NoEntityFoundException("EventStatus with id: " + id + " doesn't exist");
         });
     }
 
@@ -33,13 +32,16 @@ public class EventStatusServiceImpl implements EventStatusService {
 
     @Override
     public void deleteById(Long id) {
+        if (!eventStatusRepository.existsById(id)) {
+            throw new NoEntityFoundException("EventStatus with id: " + id + " doesn't exist");
+        }
         eventStatusRepository.deleteById(id);
     }
 
     @Override
     public EventStatus update(EventStatus eventStatus) {
         if (!eventStatusRepository.existsById(eventStatus.getId())) {
-            throw new NoEntityException("EventStatus: " + eventStatus + " doesn't exist");
+            throw new NoEntityFoundException("EventStatus: " + eventStatus + " doesn't exist");
         }
         return eventStatusRepository.save(eventStatus);
     }
