+++ src/main/java/hakaton/eventservice/service/impl/EventTypeServiceImpl.java
@@ -1,13 +1,12 @@
 package hakaton.eventservice.service.impl;
 
-import hakaton.eventservice.exception.NoEntityException;
+import java.util.List;
+import hakaton.eventservice.exception.NoEntityFoundException;
 import hakaton.eventservice.model.EventType;
 import hakaton.eventservice.repository.EventTypeRepository;
 import hakaton.eventservice.service.EventTypeService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @AllArgsConstructor
@@ -22,7 +21,7 @@ public class EventTypeServiceImpl implements EventTypeService {
     @Override
     public EventType getById(Long id) {
         return eventTypeRepository.findById(id).orElseThrow(() -> {
-            throw new NoEntityException("EventType with id: " + id + " doesn't exist");
+            throw new NoEntityFoundException("EventType with id: " + id + " doesn't exist");
         });
     }
 
@@ -33,13 +32,16 @@ public class EventTypeServiceImpl implements EventTypeService {
 
     @Override
     public void deleteById(Long id) {
+        if (!eventTypeRepository.existsById(id)) {
+            throw new NoEntityFoundException("EventType with id: " + id + " doesn't exist");
+        }
         eventTypeRepository.deleteById(id);
     }
 
     @Override
     public EventType update(EventType eventType) {
         if (!eventTypeRepository.existsById(eventType.getId())) {
-            throw new NoEntityException("EventType " + eventType + " doesn't exist");
+            throw new NoEntityFoundException("EventType " + eventType + " doesn't exist");
         }
         return eventTypeRepository.save(eventType);
     }
